requirements.txt
Django>=5.2,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=6.2,<7.0
uuid>=1.30,<2.0
